+++ Components/IRacing/Events/IRacingCompletedLap.cs
@@ -12,6 +12,7 @@ namespace Slipstream.Components.IRacing.Events
         public double LapTime { get; set; }
         public bool EstimatedLapTime { get; set; }
         public int LapsCompleted { get; set; }
+        public float? FuelLeft { get; set; }
         public float? FuelDelta { get; set; }
         public bool LocalUser { get; set; }
         public bool BestLap { get; set; }
@@ -25,6 +26,7 @@ namespace Slipstream.Components.IRacing.Events
                    LapTime == lap.LapTime &&
                    EstimatedLapTime == lap.EstimatedLapTime &&
                    LapsCompleted == lap.LapsCompleted &&
+                   FuelLeft == lap.FuelLeft &&
                    FuelDelta == lap.FuelDelta &&
                    LocalUser == lap.LocalUser &&
                    BestLap == lap.BestLap;
@@ -39,6 +41,7 @@ namespace Slipstream.Components.IRacing.Events
             hashCode = hashCode * -1521134295 + LapTime.GetHashCode();
             hashCode = hashCode * -1521134295 + EstimatedLapTime.GetHashCode();
             hashCode = hashCode * -1521134295 + LapsCompleted.GetHashCode();
+            hashCode = hashCode * -1521134295 + FuelLeft.GetHashCode();
             hashCode = hashCode * -1521134295 + FuelDelta.GetHashCode();
             hashCode = hashCode * -1521134295 + LocalUser.GetHashCode();
             hashCode = hashCode * -1521134295 + BestLap.GetHashCode();